pyproject.toml
[tool.poetry]
name = "pruebaCRUD"
version = "0.1.0"
description = ""
authors = ["Valfar Developer <valfar.developer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.1"
djangorestframework = "^3.11.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
